Move DCT into utils again
